+++ sources/putc.c
@@ -7,7 +7,7 @@ int putc(int c, FILE *stream)
     unsigned char uc;
 
     uc = (unsigned char) c;
-    ret = fwrite(&uc, 1, 1, stream);
+    ret = (int)fwrite(&uc, 1, 1, stream);
 
     return (ret < 0 ? ret : uc);
 }
